move helper functions to their own module
